recommender/: split code and tests into two files (query_helpers & single_image_query)
